add select for services in graphService and graphServiceZoom
git-svn-id: http://svn.centreon.com/oreon-dev@1591 6bcd3966-0018-0410-8128-fd23d134de7e

+++ www/include/views/graphs/graphODS/graphODSService.php
@@ -93,11 +93,21 @@ For information : contact@oreon-project.org
 		$DBRESULT2->fetchInto($svc_id);
 	}
 	
+	$DBRESULT2 =& $pearDBO->query("SELECT id, service_description  FROM index_data WHERE host_name = '".$svc_id["host_name"]."'");
+	if (PEAR::isError($DBRESULT2))
+		print "Mysql Error : ".$DBRESULT2->getDebugInfo();
+	$other_services = array(""=>"");
+	while ($DBRESULT2->fetchInto($selected_service))
+		$other_services[$selected_service["id"]] = $selected_service["service_description"];
+	$DBRESULT2->free();
+	$form->addElement('select', 'index', 'Others Services', $other_services);
+	$form->setDefaults($_GET);
+	
 	$service_id = $svc_id["service_id"];
 	$index_id = $svc_id["id"];
 	
-	$indexF =& $form->addElement('hidden', 'index');
-	$indexF->setValue($index_id);
+	//$indexF =& $form->addElement('hidden', 'index');
+	//$indexF->setValue($index_id);
 	
 	if (!$service_id) $tpl->assign('msg', $lang["no_graph_found"]); else $tpl->assign('msg', NULL);	
 	if (!isset($start) && !isset($end)){
@@ -125,16 +135,28 @@ For information : contact@oreon-project.org
 		$counter++;
 	}
 		
-	if (isset($_GET["metric"])){
-		$metrics_active =& $_GET["metric"];
+	# verify if metrics in parameter is for this index
+	
+	$metrics_active =& $_GET["metric"];
+	$pass = 0;
+	if (isset($metrics_active))
+		foreach ($metrics_active as $key => $value)
+			if (isset($metrics[$key]))
+				$pass = 1;
+	# 
+	
+	
+	if (isset($_GET["metric"]) && $pass){
 		$tpl->assign('metric_active', $metrics_active);	
 		$DBRESULT =& $pearDB->query("DELETE FROM `ods_view_details` WHERE index_id = '".$_GET["index"]."'");
 		if (PEAR::isError($DBRESULT))
 			print "Mysql Error : ".$DBRESULT->getDebugInfo();
 		foreach ($metrics_active as $key => $metric){
-			$DBRESULT =& $pearDB->query("INSERT INTO `ods_view_details` (`metric_id`, `contact_id`, `all_user`, `index_id`) VALUES ('".$key."', '".$oreon->user->user_id."', '0', '".$_GET["index"]."');");
-			if (PEAR::isError($DBRESULT))
-				print "Mysql Error : ".$DBRESULT->getDebugInfo();
+			if (isset($metrics_active[$metric["metric_id"]])){
+				$DBRESULT =& $pearDB->query("INSERT INTO `ods_view_details` (`metric_id`, `contact_id`, `all_user`, `index_id`) VALUES ('".$key."', '".$oreon->user->user_id."', '0', '".$_GET["index"]."');");
+				if (PEAR::isError($DBRESULT))
+					print "Mysql Error : ".$DBRESULT->getDebugInfo();
+			}
 		}
 	} else {
 		$DBRESULT =& $pearDB->query("SELECT metric_id FROM `ods_view_details` WHERE index_id = '".$_GET["index"]."' AND `contact_id` = '".$oreon->user->user_id."'");
@@ -161,6 +183,7 @@ For information : contact@oreon-project.org
 	$tpl->assign('host_name', $svc_id);
 
 	$tpl->assign('metrics', $metrics);
+	$tpl->assign('nb_metrics', count($metrics));
 	$tpl->assign('metrics_active', $metrics_active);
 
 	$tpl->assign('isAvl', 1);
